move registration to PrivateGateway
