+++ packages/daemon/src/sync/publicGateway/GatewayRegistrar.spec.ts
@@ -1,36 +1,17 @@
-import {
-  Certificate,
-  derSerializePublicKey,
-  GSCClient,
-  PrivateGateway,
-  PrivateNodeRegistration,
-  PrivateNodeRegistrationRequest,
-  SessionKey,
-  SessionKeyPair,
-  UnreachableResolverError,
-} from '@relaycorp/relaynet-core';
-import {
-  generateIdentityKeyPairSet,
-  generatePDACertificationPath,
-  MockGSCClient,
-  PreRegisterNodeCall,
-  RegisterNodeCall,
-} from '@relaycorp/relaynet-testing';
+import { UnreachableResolverError } from '@relaycorp/relaynet-core';
 import { Container } from 'typedi';
 
 import { Config, ConfigKey } from '../../Config';
 import { DEFAULT_PUBLIC_GATEWAY } from '../../constants';
 import { useTemporaryAppDirs } from '../../testUtils/appDirs';
-import { arrayBufferFrom } from '../../testUtils/buffer';
 import { setUpTestDBConnection } from '../../testUtils/db';
-import { mockSpy } from '../../testUtils/jest';
 import { mockLoggerToken, partialPinoLog } from '../../testUtils/logging';
 import { mockSleepSeconds } from '../../testUtils/timing';
 import { GatewayRegistrar } from './GatewayRegistrar';
-import * as gscClient from './gscClient';
-import { NonExistingAddressError, PublicGatewayProtocolError } from './errors';
-import { getPromiseRejection } from '../../testUtils/promises';
+import { NonExistingAddressError } from './errors';
 import { PrivateGatewayManager } from '../../PrivateGatewayManager';
+import { mockSpy } from '../../testUtils/jest';
+import { PrivateGateway } from '../../PrivateGateway';
 
 setUpTestDBConnection();
 
@@ -43,45 +24,14 @@ beforeEach(() => {
   registrar = Container.get(GatewayRegistrar);
 });
 
-let mockGSCClient: GSCClient | null;
-beforeEach(() => {
-  mockGSCClient = null;
-});
-
-const mockMakeGSCClient = mockSpy(jest.spyOn(gscClient, 'makeGSCClient'), () => mockGSCClient);
-
-let publicGatewayIdCertificate: Certificate;
-let publicGatewaySessionKey: SessionKey;
-let idCertificate: Certificate;
-beforeAll(async () => {
-  const certPath = await generatePDACertificationPath(await generateIdentityKeyPairSet());
-  publicGatewayIdCertificate = certPath.publicGateway;
-  idCertificate = certPath.privateGateway;
-
-  publicGatewaySessionKey = (await SessionKeyPair.generate()).sessionKey;
-});
-
 describe('register', () => {
-  const registrationAuth = arrayBufferFrom('the auth');
-  let preRegisterCall: PreRegisterNodeCall;
-  let registration: PrivateNodeRegistration;
-  let registerCall: RegisterNodeCall;
-  beforeEach(() => {
-    preRegisterCall = new PreRegisterNodeCall(registrationAuth);
-
-    registration = new PrivateNodeRegistration(
-      idCertificate,
-      publicGatewayIdCertificate,
-      publicGatewaySessionKey,
-    );
-    registerCall = new RegisterNodeCall(registration);
-
-    mockGSCClient = new MockGSCClient([preRegisterCall, registerCall]);
-  });
-
   beforeEach(async () => {
     await Container.get(PrivateGatewayManager).createCurrentIfMissing();
   });
+
+  const mockRegisterWithPublicGateway = mockSpy(
+    jest.spyOn(PrivateGateway.prototype, 'registerWithPublicGateway'),
+  );
 
   test('Registration should be skipped if already registered with new gateway', async () => {
     const config = Container.get(Config);
@@ -89,63 +39,20 @@ describe('register', () => {
 
     await registrar.register(DEFAULT_PUBLIC_GATEWAY);
 
-    expect(mockMakeGSCClient).not.toBeCalled();
-    expect(preRegisterCall.wasCalled).toBeFalsy();
+    expect(mockRegisterWithPublicGateway).not.toBeCalled();
     expect(logs).toContainEqual(
       partialPinoLog('debug', 'Skipping registration with public gateway'),
     );
   });
 
-  test('PoWeb client should complete registration with resolved address', async () => {
+  test('Registration should be completed if not already registered with peer', async () => {
     await registrar.register(DEFAULT_PUBLIC_GATEWAY);
 
-    expect(mockMakeGSCClient).toBeCalledWith(DEFAULT_PUBLIC_GATEWAY);
+    expect(mockRegisterWithPublicGateway).toBeCalledWith(DEFAULT_PUBLIC_GATEWAY);
     expect(logs).toContainEqual(
       partialPinoLog('info', 'Successfully registered with public gateway', {
         publicGatewayPublicAddress: DEFAULT_PUBLIC_GATEWAY,
-        publicGatewayPrivateAddress:
-          await publicGatewayIdCertificate.calculateSubjectPrivateAddress(),
       }),
-    );
-  });
-
-  test('PoWeb client should do pre-registration', async () => {
-    await registrar.register(DEFAULT_PUBLIC_GATEWAY);
-
-    expect(preRegisterCall.wasCalled).toBeTruthy();
-  });
-
-  test('PoWeb client should complete registration with given authorisation', async () => {
-    await registrar.register(DEFAULT_PUBLIC_GATEWAY);
-
-    expect(registerCall.wasCalled).toBeTruthy();
-    const registrationRequest = await PrivateNodeRegistrationRequest.deserialize(
-      registerCall.arguments!.pnrrSerialized,
-    );
-    await expect(derSerializePublicKey(registrationRequest.privateNodePublicKey)).resolves.toEqual(
-      await derSerializePublicKey(preRegisterCall.arguments!.nodePublicKey),
-    );
-    expect(registrationRequest.pnraSerialized).toEqual(registrationAuth);
-  });
-
-  test('Channel with public gateway should be stored', async () => {
-    const saveChannelSpy = jest.spyOn(PrivateGateway.prototype, 'savePublicGatewayChannel');
-
-    await registrar.register(DEFAULT_PUBLIC_GATEWAY);
-
-    expect(saveChannelSpy).toBeCalledWith(
-      idCertificate,
-      publicGatewayIdCertificate,
-      publicGatewaySessionKey,
-    );
-  });
-
-  test('Private address of public gateway should be stored in config', async () => {
-    await registrar.register(DEFAULT_PUBLIC_GATEWAY);
-
-    const config = Container.get(Config);
-    await expect(config.get(ConfigKey.PUBLIC_GATEWAY_PRIVATE_ADDRESS)).resolves.toEqual(
-      await publicGatewayIdCertificate.calculateSubjectPrivateAddress(),
     );
   });
 
@@ -158,53 +65,11 @@ describe('register', () => {
     );
   });
 
-  test('Error should be thrown if registration fails', async () => {
+  test('Registration errors should be propagated', async () => {
     const originalError = new Error('oh noes');
-    registerCall = new RegisterNodeCall(originalError);
-    mockGSCClient = new MockGSCClient([preRegisterCall, registerCall]);
+    mockRegisterWithPublicGateway.mockRejectedValue(originalError);
 
-    const error = await getPromiseRejection(
-      registrar.register(DEFAULT_PUBLIC_GATEWAY),
-      PublicGatewayProtocolError,
-    );
-    expect(error.message).toMatch(/^Failed to register with the public gateway:/);
-    expect(error.cause()).toBe(originalError);
-  });
-
-  test('Error should be thrown if the public gateway session key is missing', async () => {
-    registration = new PrivateNodeRegistration(idCertificate, publicGatewayIdCertificate);
-    registerCall = new RegisterNodeCall(registration);
-    mockGSCClient = new MockGSCClient([preRegisterCall, registerCall]);
-    const saveChannelSpy = jest.spyOn(PrivateGateway.prototype, 'savePublicGatewayChannel');
-
-    await expect(registrar.register(DEFAULT_PUBLIC_GATEWAY)).rejects.toThrowWithMessage(
-      PublicGatewayProtocolError,
-      'Registration is missing public gateway session key',
-    );
-
-    expect(saveChannelSpy).not.toBeCalled();
-    const config = Container.get(Config);
-    await expect(config.get(ConfigKey.PUBLIC_GATEWAY_PUBLIC_ADDRESS)).resolves.toBeNull();
-    await expect(config.get(ConfigKey.PUBLIC_GATEWAY_PRIVATE_ADDRESS)).resolves.toBeNull();
-  });
-
-  test('Error should be thrown if the channel creation is rejected', async () => {
-    registration = new PrivateNodeRegistration(
-      idCertificate,
-      idCertificate, // Invalid public gateway certificate
-      publicGatewaySessionKey,
-    );
-    registerCall = new RegisterNodeCall(registration);
-    mockGSCClient = new MockGSCClient([preRegisterCall, registerCall]);
-
-    await expect(registrar.register(DEFAULT_PUBLIC_GATEWAY)).rejects.toThrowWithMessage(
-      PublicGatewayProtocolError,
-      /^Failed to save channel with public gateway:/,
-    );
-
-    const config = Container.get(Config);
-    await expect(config.get(ConfigKey.PUBLIC_GATEWAY_PUBLIC_ADDRESS)).resolves.toBeNull();
-    await expect(config.get(ConfigKey.PUBLIC_GATEWAY_PRIVATE_ADDRESS)).resolves.toBeNull();
+    await expect(registrar.register(DEFAULT_PUBLIC_GATEWAY)).rejects.toBe(originalError);
   });
 });
